+++ src/app/native/enum.ts
@@ -1,6 +1,5 @@
 export enum Path {
   Image = 'image',
-  Detail = 'detail',
   Note = 'note',
   Knowledge = 'knowledge',
 }